cilium: add new probe package for BPF kernel feature probes
Move the LPM probe out of prefilter itself and into its own package.
