+++ pkg/datapath/prefilter/prefilter.go
@@ -21,11 +21,11 @@ import (
 	"os/exec"
 	"path"
 	"syscall"
-	"unsafe"
 
 	"github.com/cilium/cilium/pkg/bpf"
 	"github.com/cilium/cilium/pkg/lock"
 	"github.com/cilium/cilium/pkg/maps/cidrmap"
+	"github.com/cilium/cilium/pkg/probe"
 )
 
 type preFilterMapType int
@@ -278,52 +278,12 @@ func ProbePreFilter(device, mode string) error {
 	return nil
 }
 
-type probeKey struct {
-	Prefixlen uint32
-	Key       uint32
-}
-
-type probeValue struct {
-	Value uint32
-}
-
-func (p *probeKey) String() string             { return fmt.Sprintf("key=%d", p.Key) }
-func (p *probeKey) GetKeyPtr() unsafe.Pointer  { return unsafe.Pointer(p) }
-func (p *probeKey) NewValue() bpf.MapValue     { return &probeValue{} }
-func (p *probeKey) DeepCopyMapKey() bpf.MapKey { return &probeKey{p.Prefixlen, p.Key} }
-
-func (p *probeValue) String() string                 { return fmt.Sprintf("value=%d", p.Value) }
-func (p *probeValue) GetValuePtr() unsafe.Pointer    { return unsafe.Pointer(p) }
-func (p *probeValue) DeepCopyMapValue() bpf.MapValue { return &probeValue{p.Value} }
-
-func ProbePreFilterDyn() bool {
-	m := bpf.NewMap("cilium_test", bpf.MapTypeLPMTrie,
-		&probeKey{}, int(unsafe.Sizeof(probeKey{})),
-		&probeValue{}, int(unsafe.Sizeof(probeValue{})),
-		1, bpf.BPF_F_NO_PREALLOC, 0, bpf.ConvertKeyValue).WithCache()
-	_, err := m.OpenOrCreateUnpinned()
-	defer m.Close()
-	if err != nil {
-		return false
-	}
-	err = bpf.UpdateElement(m.GetFd(), unsafe.Pointer(&probeKey{}),
-		unsafe.Pointer(&probeValue{}), bpf.BPF_ANY)
-	if err != nil {
-		return false
-	}
-	err = bpf.GetNextKey(m.GetFd(), nil, unsafe.Pointer(&probeKey{}))
-	if err != nil {
-		return false
-	}
-	return true
-}
-
 // NewPreFilter returns prefilter handle
 func NewPreFilter() (*PreFilter, error) {
-	haveDyn := ProbePreFilterDyn()
+	haveLPM := probe.HaveFullLPM()
 	c := preFilterConfig{
-		dyn4Enabled: haveDyn,
-		dyn6Enabled: haveDyn,
+		dyn4Enabled: haveLPM,
+		dyn6Enabled: haveLPM,
 		fix4Enabled: true,
 		fix6Enabled: true,
 	}
